dev-next : RsWeb - Format Styles  destructor deleted also a static member
It is a bug because atm FormatStyles are destroyed for each run. In the previous version, before porting to net core,  Formatters were stored in Application bag, so their destructor was called once at the end of application lifecycle. It's no longer valid. NEED TO BE IMPROVED AND FORMATTERS STORED TO LIVE ACROSS REPORTS EXECUTION.